Bug fixes and performance improvements
Fixed a problem in Topography where I misinterpreted what height was
referring to. Fixed a computation error that slowed the model down
significantly. If every cell in the model contains water, it takes
somewhere between 3.8-4.3 seconds per time step. Now I'm trying to fix
an issue where the total water in the system is going crazy.