WaterFlow tweaks
WaterFlow can now simulate 7 months in about 5 minutes with an accuracy
of + or - 1 milliliter of water (over the entire 7 month period). The
time step is huge (950 seconds ~ 16 minutes), but since the water flows
so slowly, it stays accurate. Lastly, I added in a quick way to suppress
my program's output so you don't have to go through and comment
everything.