Updated GUI, CreateFarmGUI, and Local_Control
WaterProjectGUI now takes a name when it is created. That way we can
find the buyer and seller in a list of GUIs. And give the correct money
and quantity to them. Create Farm GUI now checks what the x,y,z cell
count is for the farm and updates those values to the sliders on the
GUI. Local Control I just added a check that if the seller is the buyer
then the order is removed. So that is how a seller can cancel their
offers.