GUI Change
When GUI is created, it first opens up a input dialog box, where the
user types in a name. That name is saved for the GUI, and "Name" class
is called to create a Name using the given user input name. *still need
to send this "Name" out to the server like we discussed.